applications/rainbowcake_video_processing/original/requirements.txt
imageio-ffmpeg==0.4.9
ffmpeg-python==0.2.0